feat: 최종 승패를 출력하는 기능
